backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
python-json-logger==2.0.7